+++ routes/web.php
@@ -14,10 +14,9 @@ use Illuminate\Support\Facades\Route;
 */
 
 Route::get('/', 'DashboardController@index')->name('home');
-Route::get('/koleksi', 'DashboardController@koleksi')->name('koleksi');
-Route::get('/pinjaman', 'DashboardController@pinjaman')->name('pinjaman');
-Route::get('/kunjungan', 'DashboardController@kunjungan')->name('kunjungan');
-Route::get('/graph-data', 'DashboardController@graph_data')->name('graph-data');
-Route::get('/graph-data', 'DashboardController@graph_data')->name('graph-data');
-Route::get('/graph-kunjungan', 'DashboardController@graph_kunjungan')->name('graph-kunjungan');
-Route::get('/graph-pinjaman', 'DashboardController@graph_pinjaman')->name('graph-pinjaman');
+// Auth::routes();
+// Route::group(['middleware' => ['auth']], function () {
+    Route::get('koleksi', 'DashboardController@koleksi')->name('koleksi');
+    Route::get('pinjaman', 'DashboardController@pinjaman')->name('pinjaman');
+    Route::get('kunjungan', 'DashboardController@kunjungan')->name('kunjungan');
+// });
